Fix the way of getting Instagram Account ID
The way Facebook allow to get Instagram Account ID has changed. Now for
the user to be able to use with this lib their Instagram Accounts has to
meet to requirements:
1. Be an Instagram Business Account
2. Be this Instagram Business Account linked to a Facebook page.

So this PR try to fix this lib by adding 3 new endpoints and removing
the old `/me` which is not used anymore.
The new endpoints are:
1. `/me/accounts` to get a list of authorized FB pages by the user (accessToken)
2. /${facebook-page-id} to get the linked Instagram Business account of that page.
3. /${instagram-id} to get Instagram account info like `username` or
   `profle_picture`

Now is when we can start publishing content on behalve of the user.

As bonus we introduce another enpdpoint to get a long-lived access token
for the user is using your app. This is helpful when the app you're
developing is meant to be used for people outside your company.